testproj/requirements.txt
django
djangorestframework
jsbeautifier
-e ..
